+++ src/follow_cursor.ts
@@ -10,6 +10,56 @@ import { ensure_websocket_server } from './rpe'
 const logger = get_logger()
 const last_warps = new Map<number, string>()
 
+interface SyncEditorWithRenpyOptions {
+	/** absolute path to the file */
+	path: string
+	/** path relative from the game folder (e.g. `script.rpy`) */
+	relative_path: string
+	/** 0-indexed line number */
+	line: number
+}
+
+export async function sync_editor_with_renpy({
+	path,
+	relative_path,
+	line,
+}: SyncEditorWithRenpyOptions): Promise<void> {
+	if (
+		!["Ren'Py updates Visual Studio Code", 'Update both'].includes(
+			get_config('followCursorMode')
+		)
+	)
+		return
+
+	// prevent feedback loop with warp to cursor
+	//
+	// TODO: this will still happen if renpy warps to a different line
+	// than the one requested.
+	//
+	// last_warp_spec = `${relative_path}:${line}`
+
+	const doc = await vscode.workspace.openTextDocument(path)
+	await vscode.window.showTextDocument(doc)
+	const editor = vscode.window.activeTextEditor
+
+	if (!editor) {
+		logger.warn('no active text editor')
+		return
+	}
+
+	// if the cursor is already on the correct line, don't munge it
+	if (editor.selection.start.line !== line) {
+		logger.debug(`syncing editor to ${relative_path}:${line}`)
+
+		const end_of_line = editor.document.lineAt(line).range.end.character
+		const pos = new vscode.Position(line, end_of_line)
+		const selection = new vscode.Selection(pos, pos)
+
+		editor.selection = selection
+		editor.revealRange(selection)
+	}
+}
+
 export async function warp_renpy_to_cursor(rp: RenpyProcess) {
 	const editor = vscode.window.activeTextEditor
 
@@ -85,14 +135,6 @@ export class FollowCursor {
 	async enable() {
 		if (!this.pm) throw new Error('no ProcessManager in FollowCursor')
 
-		if (!get_config('renpyExtensionsEnabled')) {
-			vscode.window.showErrorMessage(
-				"Follow cursor only works with Ren'Py extensions enabled.",
-				'OK'
-			)
-			return
-		}
-
 		if (this.pm.length > 1) {
 			vscode.window.showErrorMessage(
 				"Can't follow cursor with multiple open processes",
@@ -105,6 +147,14 @@ export class FollowCursor {
 		if (process === undefined) {
 			vscode.window.showErrorMessage(
 				"Ren'Py not running. Cannot follow cursor.",
+				'OK'
+			)
+			return
+		}
+
+		if (!get_config('renpyExtensionsEnabled')) {
+			vscode.window.showErrorMessage(
+				"Follow cursor only works with Ren'Py extensions enabled.",
 				'OK'
 			)
 			return