+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
--   Don't link to specific settings (it doesn't work very well)
 -   Migrate source code to TypeScript
+-   Don't link to specific settings (it doesn't work very well)
+-   You can no longer start the extension with `renpyWarp.toggleFollowCursor`
+-   Say what process an output line was from
 
 ## 1.2.0 - 2024-07-06